requirements.txt
sentence-transformers
faiss-cpu
rank-bm25
numpy
requests
tqdm
